requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
Werkzeug==3.0.1
# CrewAI and LLM dependencies
crewai>=0.28.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-groq>=0.1.0
groq>=0.4.0
python-dotenv>=1.0.0
litellm>=1.0.0

# PDF processing
PyPDF2>=3.0.0
# Test dependencies
reportlab==4.0.7
